requirements.txt
Flask>=2.2.0
flask-cors
google-cloud-datastore
google-cloud-storage
requests
gunicorn==20.1.0